Rename composition StackedBarChart -> BarChart
It's not necessarily a stacked bar chart, so a more generic name seems appropriate.
