+++ src/ui/compositions/bar-chart/README.md
