+++ src/ui/compositions/bar-chart/demo/index.html
