+++ src/ui/compositions/bar-chart/index.js
@@ -1,0 +1,1 @@
+export { default as default } from './src/bar-chart';
