+++ src/ui/compositions/bar-chart/src/style.css
