--- src/ui/compositions/stacked/index.js
@@ -1,1 +1,0 @@
-export { default as default } from './src/stacked';
